fix(participate): Fix bug in has-not-participate-validator
I have fixed a bug in the has-not-participate-validator related to the participate feature Previously, the validator allowed a user to participate in a poll multiple times if they had previously participated but chose a different option. However, the rule is that each user can only participate in a poll once. This bug has been fixed to ensure that the validator correctly enforces this rule.